requirements.txt
watchdog 
pytesseract 
pillow 
pyperclip
InquirerPy